requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
streamlit>=1.30.0
Pillow>=9.0.0
numpy>=1.24.0